refactor(typed-store): get rid of run-time enum dispatch pattern in favour of static dispatch via traits
