+++ crates/iota-core/src/jsonrpc_index.rs
@@ -45,10 +45,11 @@ use tracing::{debug, info, trace};
 use typed_store::{
     DBMapUtils, TypedStoreError,
     rocks::{
-        DBBatch, DBMap, DBMapTableConfigMap, DBOptions, MetricConf, default_db_options,
+        DBMapTableConfigMap, DBOptions, default_db_options,
         read_size_from_env,
     },
     rocksdb::compaction_filter::Decision,
+    store::{DBBatch, DBMap, MetricConf},
     traits::{Map, TableSummary, TypedStoreDebug},
 };
