+++ crates/iota-storage/src/write_path_pending_tx_log.rs
@@ -21,7 +21,7 @@ use iota_types::{
 use tracing::instrument;
 use typed_store::{
     DBMapUtils,
-    rocks::{DBMap, MetricConf},
+    store::{DBMap, MetricConf},
     traits::{Map, TableSummary, TypedStoreDebug},
 };
