+++ crates/typed-store-derive/src/lib.rs
@@ -410,8 +410,8 @@ pub fn derive_dbmap_utils_general(input: TokenStream) -> TokenStream {
             pub fn open_tables_impl(
                 path: std::path::PathBuf,
                 as_secondary_with_path: Option<std::path::PathBuf>,
-                metric_conf: typed_store::rocks::MetricConf,
-                global_db_options_override: Option<typed_store::rocksdb::Options>,
+                metric_conf: typed_store::store::MetricConf,
+                global_db_options_override: Option<typed_store::rocks::Options>,
                 tables_db_options_override: Option<typed_store::rocks::DBMapTableConfigMap>,
             ) -> Self {
                 let path = &path;
@@ -500,8 +500,8 @@ pub fn derive_dbmap_utils_general(input: TokenStream) -> TokenStream {
             #[expect(unused_parens)]
             pub fn open_tables_read_write(
                 path: std::path::PathBuf,
-                metric_conf: typed_store::rocks::MetricConf,
-                global_db_options_override: Option<typed_store::rocksdb::Options>,
+                metric_conf: typed_store::store::MetricConf,
+                global_db_options_override: Option<typed_store::rocks::Options>,
                 tables_db_options_override: Option<typed_store::rocks::DBMapTableConfigMap>
             ) -> Self {
                 let inner = #intermediate_db_map_struct_name::open_tables_impl(path, None, metric_conf, global_db_options_override, tables_db_options_override);
@@ -524,7 +524,7 @@ pub fn derive_dbmap_utils_general(input: TokenStream) -> TokenStream {
                 primary_path: std::path::PathBuf,
                 with_secondary_path: Option<std::path::PathBuf>,
                 global_db_options_override: Option<typed_store::rocksdb::Options>,
-                metric_conf: typed_store::rocks::MetricConf,
+                metric_conf: typed_store::store::MetricConf,
                 ) -> #secondary_db_map_struct_name #generics {
                 #secondary_db_map_struct_name::open_tables_read_only(primary_path, with_secondary_path, metric_conf, global_db_options_override)
             }
@@ -552,8 +552,8 @@ pub fn derive_dbmap_utils_general(input: TokenStream) -> TokenStream {
             pub fn open_tables_read_only(
                 primary_path: std::path::PathBuf,
                 with_secondary_path: Option<std::path::PathBuf>,
-                metric_conf: typed_store::rocks::MetricConf,
-                global_db_options_override: Option<typed_store::rocksdb::Options>,
+                metric_conf: typed_store::store::MetricConf,
+                global_db_options_override: Option<typed_store::rocks::Options>,
             ) -> Self {
                 let inner = match with_secondary_path {
                     Some(q) => #intermediate_db_map_struct_name::open_tables_impl(primary_path, Some(q), metric_conf, global_db_options_override, None),
