+++ crates/typed-store/src/rocks/mod.rs
@@ -1,50 +1,35 @@
 // Copyright (c) Mysten Labs, Inc.
-// Modifications Copyright (c) 2024 IOTA Stiftung
+// Modifications Copyright (c) 2026 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod errors;
-pub mod memstore;
-pub(crate) mod safe_iter;
+pub mod metrics;
+pub mod safe_iter;
 
 use std::{
-    borrow::Borrow,
-    collections::{BTreeMap, HashSet},
+    collections::HashSet,
     env,
     ffi::CStr,
-    marker::PhantomData,
-    ops::{Bound, Deref, RangeBounds},
     path::{Path, PathBuf},
     sync::Arc,
     time::Duration,
 };
 
 use backoff::backoff::Backoff;
-use bincode::Options;
 use iota_macros::{fail_point, nondeterministic};
-use prometheus::{Histogram, HistogramTimer};
 use rocksdb::{
     AsColumnFamilyRef, BlockBasedOptions, Cache, ColumnFamilyDescriptor, DBPinnableSlice,
-    DBWithThreadMode, Error, LiveFile, MultiThreaded, ReadOptions, WriteBatch, WriteOptions,
-    checkpoint::Checkpoint, properties, properties::num_files_at_level,
+    DBWithThreadMode, LiveFile, MultiThreaded, checkpoint::Checkpoint, properties,
+    properties::num_files_at_level,
 };
-use serde::{Serialize, de::DeserializeOwned};
+pub use rocksdb::Options;
 use tap::TapFallible;
-use tokio::sync::oneshot;
-use tracing::{debug, error, info, instrument, warn};
+use tracing::{info, instrument, warn};
 use typed_store_error::TypedStoreError;
 
-use crate::{
-    metrics::{DBMetrics, RocksDBPerfContext, SamplingInterval},
-    rocks::{
-        errors::{
-            typed_store_err_from_bcs_err, typed_store_err_from_bincode_err,
-            typed_store_err_from_rocks_err,
-        },
-        memstore::{InMemoryBatch, InMemoryDB},
-        safe_iter::{SafeIter, SafeRevIter},
-    },
-    traits::{Map, TableSummary},
-};
+use crate::engine::{RawDbIterator, StorageEngine};
+use errors::typed_store_err_from_rocks_err;
+use metrics::DBMetrics;
 
 // Write buffer size per RocksDB instance can be set via the env var below.
 // If the env var is not set, use the default value in MiB.
@@ -83,6 +68,100 @@ const DB_CORRUPTED_KEY: &[u8] = b"db_corrupted";
 #[cfg(test)]
 mod tests;
 
+use crate::store::{DBMap, Database, MetricConf};
+
+// ---------------------------------------------------------------------------
+// ReadWriteOptions — RocksDB-specific semantic read/write configuration
+// ---------------------------------------------------------------------------
+
+/// Semantic options for a `DBMap<K, V, RocksDB>` instance.
+///
+/// These are converted into backend-specific `RocksReadOptions` /
+/// `RocksWriteOptions` when a column-family map is opened via `reopen`.
+#[derive(Clone, Debug)]
+pub struct ReadWriteOptions {
+    pub ignore_range_deletions: bool,
+    pub(crate) sync_to_disk: bool,
+}
+
+impl ReadWriteOptions {
+    pub fn set_ignore_range_deletions(mut self, ignore: bool) -> Self {
+        self.ignore_range_deletions = ignore;
+        self
+    }
+}
+
+impl Default for ReadWriteOptions {
+    fn default() -> Self {
+        Self {
+            ignore_range_deletions: true,
+            sync_to_disk: std::env::var("IOTA_DB_SYNC_TO_DISK").is_ok_and(|v| v != "0"),
+        }
+    }
+}
+
+// ---------------------------------------------------------------------------
+// RocksReadOptions / RocksWriteOptions — cloneable wrappers around rocksdb
+// option types used as StorageEngine associated types
+// ---------------------------------------------------------------------------
+
+/// Cloneable read-option state for the RocksDB backend.
+///
+/// Stores the semantic flags and iteration bounds that are translated into a
+/// `rocksdb::ReadOptions` just-in-time inside each `StorageEngine` method.
+#[derive(Clone, Debug)]
+pub struct RocksReadOptions {
+    pub(crate) ignore_range_deletions: bool,
+    pub(crate) lower_bound: Option<Vec<u8>>,
+    pub(crate) upper_bound: Option<Vec<u8>>,
+}
+
+impl Default for RocksReadOptions {
+    fn default() -> Self {
+        Self {
+            ignore_range_deletions: true,
+            lower_bound: None,
+            upper_bound: None,
+        }
+    }
+}
+
+impl RocksReadOptions {
+    fn to_rocksdb(&self) -> rocksdb::ReadOptions {
+        let mut opts = rocksdb::ReadOptions::default();
+        opts.set_ignore_range_deletions(self.ignore_range_deletions);
+        if let Some(lb) = self.lower_bound.clone() {
+            opts.set_iterate_lower_bound(lb);
+        }
+        if let Some(ub) = self.upper_bound.clone() {
+            opts.set_iterate_upper_bound(ub);
+        }
+        opts
+    }
+}
+
+/// Cloneable write-option state for the RocksDB backend.
+#[derive(Clone, Debug)]
+pub struct RocksWriteOptions {
+    pub(crate) sync_to_disk: bool,
+}
+
+impl Default for RocksWriteOptions {
+    fn default() -> Self {
+        Self {
+            sync_to_disk: std::env::var("IOTA_DB_SYNC_TO_DISK").is_ok_and(|v| v != "0"),
+        }
+    }
+}
+
+impl RocksWriteOptions {
+    fn to_rocksdb(&self) -> rocksdb::WriteOptions {
+        let mut opts = rocksdb::WriteOptions::default();
+        opts.set_sync(self.sync_to_disk);
+        opts
+    }
+}
+
 /// A helper macro to reopen multiple column families. The macro returns
 /// a tuple of DBMap structs in the same order that the column families
 /// are defined.
@@ -103,7 +182,6 @@ mod tests;
 /// use tempfile::tempdir;
 /// use prometheus::Registry;
 /// use std::sync::Arc;
-/// use typed_store::metrics::DBMetrics;
 /// use core::fmt::Error;
 ///
 /// #[tokio::main]
@@ -131,9 +209,59 @@ macro_rules! reopen {
     };
 }
 
+// ---------------------------------------------------------------------------
+// RocksDB backend
+// ---------------------------------------------------------------------------
+
+pub type RawIter<'a> =
+    rocksdb::DBRawIteratorWithThreadMode<'a, DBWithThreadMode<MultiThreaded>>;
+
+impl RawDbIterator for RawIter<'_> {
+    fn seek_to_first(&mut self) {
+        self.seek_to_first();
+    }
+
+    fn seek_to_last(&mut self) {
+        self.seek_to_last();
+    }
+
+    fn seek(&mut self, key: &[u8]) {
+        self.seek(key);
+    }
+
+    fn seek_for_prev(&mut self, key: &[u8]) {
+        self.seek_for_prev(key);
+    }
+
+    fn valid(&self) -> bool {
+        self.valid()
+    }
+
+    fn key(&self) -> Option<&[u8]> {
+        self.key()
+    }
+
+    fn value(&self) -> Option<&[u8]> {
+        self.value()
+    }
+
+    fn next(&mut self) {
+        self.next();
+    }
+
+    fn prev(&mut self) {
+        self.prev();
+    }
+
+    fn status(&self) -> Result<(), TypedStoreError> {
+        self.status()
+            .map_err(|e| TypedStoreError::RocksDB(format!("{e}")))
+    }
+}
+
 #[derive(Debug)]
 pub struct RocksDB {
-    pub underlying: rocksdb::DBWithThreadMode<MultiThreaded>,
+    pub(crate) underlying: rocksdb::DBWithThreadMode<MultiThreaded>,
 }
 
 impl Drop for RocksDB {
@@ -142,370 +270,510 @@ impl Drop for RocksDB {
     }
 }
 
-#[derive(Clone)]
-pub enum ColumnFamily {
-    Rocks(String),
-    InMemory(String),
-}
+impl RocksDB {
+    /// Retrieve a column family handle, panicking if it does not exist.
+    fn rocks_cf(&self, cf_name: &str) -> Arc<rocksdb::BoundColumnFamily<'_>> {
+        self.underlying
+            .cf_handle(cf_name)
+            .expect("Map-keying column family should have been checked at DB creation")
+    }
 
-impl std::fmt::Debug for ColumnFamily {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ColumnFamily::Rocks(name) => write!(f, "RocksDB cf: {}", name),
-            ColumnFamily::InMemory(name) => write!(f, "InMemory cf: {}", name),
+    fn get_int_property(
+        &self,
+        cf: &impl AsColumnFamilyRef,
+        property_name: &std::ffi::CStr,
+    ) -> Result<i64, TypedStoreError> {
+        match self.underlying.property_int_value_cf(cf, property_name) {
+            Ok(Some(value)) => Ok(value.min(i64::MAX as u64).try_into().unwrap_or_default()),
+            Ok(None) => Ok(0),
+            Err(e) => Err(TypedStoreError::RocksDB(e.into_string())),
         }
     }
 }
 
-impl ColumnFamily {
-    fn rocks_cf<'a>(&self, rocks_db: &'a RocksDB) -> Arc<rocksdb::BoundColumnFamily<'a>> {
-        match &self {
-            ColumnFamily::Rocks(name) => rocks_db
-                .underlying
-                .cf_handle(name)
-                .expect("Map-keying column family should have been checked at DB creation"),
-            _ => unreachable!("invariant is checked by the caller"),
-        }
+impl StorageEngine for RocksDB {
+    type Batch = rocksdb::WriteBatch;
+    type GetValue<'a>
+        = DBPinnableSlice<'a>
+    where
+        Self: 'a;
+    type ReadOptions = RocksReadOptions;
+    type WriteOptions = RocksWriteOptions;
+    type CfOptions = rocksdb::Options;
+    type RawIter<'a>
+        = RawIter<'a>
+    where
+        Self: 'a;
+    type Metrics = DBMetrics;
+
+    fn get_metrics() -> Arc<DBMetrics> {
+        DBMetrics::get().clone()
     }
-}
 
-pub enum Storage {
-    Rocks(RocksDB),
-    InMemory(InMemoryDB),
-}
-
-impl std::fmt::Debug for Storage {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Storage::Rocks(db) => write!(f, "RocksDB Storage {:?}", db),
-            Storage::InMemory(db) => write!(f, "InMemoryDB Storage {:?}", db),
-        }
+    fn set_iter_lower_bound(opts: &mut RocksReadOptions, bound: Vec<u8>) {
+        opts.lower_bound = Some(bound);
     }
-}
 
-#[derive(Debug)]
-pub struct Database {
-    storage: Storage,
-    metric_conf: MetricConf,
-}
-
-impl Drop for Database {
-    fn drop(&mut self) {
-        DBMetrics::get().decrement_num_active_dbs(&self.metric_conf.db_name);
+    fn set_iter_upper_bound(opts: &mut RocksReadOptions, bound: Vec<u8>) {
+        opts.upper_bound = Some(bound);
     }
-}
 
-enum GetResult<'a> {
-    Rocks(DBPinnableSlice<'a>),
-    InMemory(Vec<u8>),
-}
-
-impl Deref for GetResult<'_> {
-    type Target = [u8];
-    fn deref(&self) -> &[u8] {
-        match self {
-            GetResult::Rocks(d) => d.deref(),
-            GetResult::InMemory(d) => d.deref(),
-        }
-    }
-}
-
-impl Database {
-    fn new(storage: Storage, metric_conf: MetricConf) -> Self {
-        DBMetrics::get().increment_num_active_dbs(&metric_conf.db_name);
-        Self {
-            storage,
-            metric_conf,
-        }
+    fn default_cf_options() -> rocksdb::Options {
+        default_db_options().options
     }
 
     fn get<K: AsRef<[u8]>>(
         &self,
-        cf: &ColumnFamily,
+        cf_name: &str,
         key: K,
-        readopts: &ReadOptions,
-    ) -> Result<Option<GetResult<'_>>, TypedStoreError> {
-        match (&self.storage, cf) {
-            (Storage::Rocks(db), ColumnFamily::Rocks(_)) => Ok(db
-                .underlying
-                .get_pinned_cf_opt(&cf.rocks_cf(db), key, readopts)
-                .map_err(typed_store_err_from_rocks_err)?
-                .map(GetResult::Rocks)),
-            (Storage::InMemory(db), ColumnFamily::InMemory(cf_name)) => {
-                Ok(db.get(cf_name, key).map(GetResult::InMemory))
-            }
-
-            _ => Err(TypedStoreError::RocksDB(
-                "typed store invariant violation".to_string(),
-            )),
-        }
+        readopts: &RocksReadOptions,
+    ) -> Result<Option<DBPinnableSlice<'_>>, TypedStoreError> {
+        self.underlying
+            .get_pinned_cf_opt(&self.rocks_cf(cf_name), key, &readopts.to_rocksdb())
+            .map_err(typed_store_err_from_rocks_err)
     }
 
     fn multi_get<I, K>(
         &self,
-        cf: &ColumnFamily,
+        cf_name: &str,
         keys: I,
-        readopts: &ReadOptions,
-    ) -> Vec<Result<Option<GetResult<'_>>, TypedStoreError>>
+        readopts: &RocksReadOptions,
+    ) -> Vec<Result<Option<DBPinnableSlice<'_>>, TypedStoreError>>
     where
         I: IntoIterator<Item = K>,
         K: AsRef<[u8]>,
     {
-        match (&self.storage, cf) {
-            (Storage::Rocks(db), ColumnFamily::Rocks(_)) => {
-                let res = db.underlying.batched_multi_get_cf_opt(
-                    &cf.rocks_cf(db),
-                    keys,
-                    // sorted_input
-                    false,
-                    readopts,
-                );
-                res.into_iter()
-                    .map(|r| {
-                        r.map_err(typed_store_err_from_rocks_err)
-                            .map(|item| item.map(GetResult::Rocks))
-                    })
-                    .collect()
-            }
-            (Storage::InMemory(db), ColumnFamily::InMemory(cf_name)) => db
-                .multi_get(cf_name, keys)
-                .into_iter()
-                .map(|r| Ok(r.map(GetResult::InMemory)))
-                .collect(),
-            _ => unreachable!("typed store invariant violation"),
-        }
+        self.underlying
+            .batched_multi_get_cf_opt(&self.rocks_cf(cf_name), keys, false, &readopts.to_rocksdb())
+            .into_iter()
+            .map(|r| r.map_err(typed_store_err_from_rocks_err))
+            .collect()
     }
 
-    pub fn create_cf<N: AsRef<str>>(
+    fn put(
         &self,
-        name: N,
-        opts: &rocksdb::Options,
-    ) -> Result<(), rocksdb::Error> {
-        match &self.storage {
-            Storage::Rocks(db) => db.underlying.create_cf(name, opts),
-            Storage::InMemory(_) => Ok(()),
-        }
-    }
-
-    pub fn cf_handle(&self, name: &str) -> Option<()> {
-        match &self.storage {
-            Storage::Rocks(db) => db.underlying.cf_handle(name).map(|_| ()),
-            Storage::InMemory(db) => db.has_cf(name).then_some(()),
-        }
-    }
-
-    pub fn drop_cf(&self, name: &str) -> Result<(), rocksdb::Error> {
-        match &self.storage {
-            Storage::Rocks(db) => db.underlying.drop_cf(name),
-            Storage::InMemory(db) => {
-                db.drop_cf(name);
-                Ok(())
-            }
-        }
-    }
-
-    fn delete_cf<K: AsRef<[u8]>>(
-        &self,
-        cf: &ColumnFamily,
-        key: K,
-        writeopts: &WriteOptions,
-    ) -> Result<(), TypedStoreError> {
-        fail_point!("delete-cf-before");
-        let ret = match (&self.storage, cf) {
-            (Storage::Rocks(db), ColumnFamily::Rocks(_)) => db
-                .underlying
-                .delete_cf_opt(&cf.rocks_cf(db), key, writeopts)
-                .map_err(typed_store_err_from_rocks_err),
-            (Storage::InMemory(db), ColumnFamily::InMemory(cf_name)) => {
-                db.delete(cf_name, key.as_ref());
-                Ok(())
-            }
-            _ => Err(TypedStoreError::RocksDB(
-                "typed store invariant violation".to_string(),
-            )),
-        };
-        fail_point!("delete-cf-after");
-        #[allow(clippy::let_and_return)]
-        ret
-    }
-
-    pub fn path_for_pruning(&self) -> &Path {
-        match &self.storage {
-            Storage::Rocks(rocks) => rocks.underlying.path(),
-            _ => unimplemented!("method is only supported for rocksdb backend"),
-        }
-    }
-
-    fn put_cf(
-        &self,
-        cf: &ColumnFamily,
+        cf_name: &str,
         key: Vec<u8>,
         value: Vec<u8>,
-        writeopts: &WriteOptions,
+        writeopts: &RocksWriteOptions,
     ) -> Result<(), TypedStoreError> {
         fail_point!("put-cf-before");
-        let ret = match (&self.storage, cf) {
-            (Storage::Rocks(db), ColumnFamily::Rocks(_)) => db
-                .underlying
-                .put_cf_opt(&cf.rocks_cf(db), key, value, writeopts)
-                .map_err(typed_store_err_from_rocks_err),
-            (Storage::InMemory(db), ColumnFamily::InMemory(cf_name)) => {
-                db.put(cf_name, key, value);
-                Ok(())
-            }
-            _ => Err(TypedStoreError::RocksDB(
-                "typed store invariant violation".to_string(),
-            )),
-        };
+        let ret = self
+            .underlying
+            .put_cf_opt(&self.rocks_cf(cf_name), key, value, &writeopts.to_rocksdb())
+            .map_err(typed_store_err_from_rocks_err);
         fail_point!("put-cf-after");
         #[allow(clippy::let_and_return)]
         ret
     }
 
-    pub fn key_may_exist_cf<K: AsRef<[u8]>>(
+    fn delete(
         &self,
         cf_name: &str,
-        key: K,
-        readopts: &ReadOptions,
-    ) -> bool {
-        match &self.storage {
-            // [`rocksdb::DBWithThreadMode::key_may_exist_cf`] can have false positives,
-            // but no false negatives. We use it to short-circuit the absent case
-            Storage::Rocks(rocks) => {
-                rocks
-                    .underlying
-                    .key_may_exist_cf_opt(&rocks_cf(rocks, cf_name), key, readopts)
-            }
-            _ => true,
-        }
-    }
-
-    pub fn write(
-        &self,
-        batch: StorageWriteBatch,
-        writeopts: &WriteOptions,
+        key: &[u8],
+        writeopts: &RocksWriteOptions,
     ) -> Result<(), TypedStoreError> {
-        fail_point!("batch-write-before");
-        let ret = match (&self.storage, batch) {
-            (Storage::Rocks(rocks), StorageWriteBatch::Rocks(batch)) => rocks
-                .underlying
-                .write_opt(batch, writeopts)
-                .map_err(typed_store_err_from_rocks_err),
-            (Storage::InMemory(db), StorageWriteBatch::InMemory(batch)) => {
-                db.write(batch);
-                Ok(())
-            }
-            _ => Err(TypedStoreError::RocksDB(
-                "using invalid batch type for the database".to_string(),
-            )),
-        };
-        fail_point!("batch-write-after");
-
+        fail_point!("delete-cf-before");
+        let ret = self
+            .underlying
+            .delete_cf_opt(&self.rocks_cf(cf_name), key, &writeopts.to_rocksdb())
+            .map_err(typed_store_err_from_rocks_err);
+        fail_point!("delete-cf-after");
         #[allow(clippy::let_and_return)]
         ret
     }
 
-    fn raw_iterator_cf<'a: 'b, 'b>(&'a self, cf_name: &str, readopts: ReadOptions) -> RawIter<'b> {
-        match &self.storage {
-            Storage::Rocks(rocksdb) => rocksdb
-                .underlying
-                .raw_iterator_cf_opt(&rocks_cf(rocksdb, cf_name), readopts),
-            _ => unimplemented!("iterators not yet supported for other backends"),
-        }
+    fn new_batch(&self) -> rocksdb::WriteBatch {
+        rocksdb::WriteBatch::default()
     }
 
-    pub fn compact_range_cf<K: AsRef<[u8]>>(
+    fn batch_put(
         &self,
+        batch: &mut rocksdb::WriteBatch,
         cf_name: &str,
-        start: Option<K>,
-        end: Option<K>,
-    ) {
-        if let Storage::Rocks(rocksdb) = &self.storage {
-            rocksdb
-                .underlying
-                .compact_range_cf(&rocks_cf(rocksdb, cf_name), start, end);
-        }
-    }
-
-    pub fn flush(&self) -> Result<(), TypedStoreError> {
-        match &self.storage {
-            Storage::Rocks(db) => db
-                .underlying
-                .flush()
-                .map_err(typed_store_err_from_rocks_err),
-            Storage::InMemory(_) => Ok(()),
-        }
-    }
-
-    pub fn checkpoint(&self, path: &Path) -> Result<(), TypedStoreError> {
-        // TODO: implement for other storage types
-        if let Storage::Rocks(rocks) = &self.storage {
-            let checkpoint =
-                Checkpoint::new(&rocks.underlying).map_err(typed_store_err_from_rocks_err)?;
-            checkpoint
-                .create_checkpoint(path)
-                .map_err(|e| TypedStoreError::RocksDB(e.to_string()))?;
-        }
+        key: Vec<u8>,
+        value: Vec<u8>,
+    ) -> Result<(), TypedStoreError> {
+        batch.put_cf(&self.rocks_cf(cf_name), key, value);
         Ok(())
     }
 
-    pub fn get_sampling_interval(&self) -> SamplingInterval {
-        self.metric_conf.read_sample_interval.new_from_self()
+    fn batch_delete(
+        &self,
+        batch: &mut rocksdb::WriteBatch,
+        cf_name: &str,
+        key: Vec<u8>,
+    ) -> Result<(), TypedStoreError> {
+        batch.delete_cf(&self.rocks_cf(cf_name), key);
+        Ok(())
     }
 
-    pub fn multiget_sampling_interval(&self) -> SamplingInterval {
-        self.metric_conf.read_sample_interval.new_from_self()
+    fn batch_delete_range(
+        &self,
+        batch: &mut rocksdb::WriteBatch,
+        cf_name: &str,
+        from: Vec<u8>,
+        to: Vec<u8>,
+    ) -> Result<(), TypedStoreError> {
+        batch.delete_range_cf(&self.rocks_cf(cf_name), from, to);
+        Ok(())
     }
 
-    pub fn write_sampling_interval(&self) -> SamplingInterval {
-        self.metric_conf.write_sample_interval.new_from_self()
+    fn batch_size_in_bytes(&self, batch: &rocksdb::WriteBatch) -> usize {
+        batch.size_in_bytes()
     }
 
-    pub fn iter_sampling_interval(&self) -> SamplingInterval {
-        self.metric_conf.iter_sample_interval.new_from_self()
-    }
-
-    fn db_name(&self) -> String {
-        let name = &self.metric_conf.db_name;
-        if name.is_empty() {
-            "default".to_string()
-        } else {
-            name.clone()
-        }
-    }
-
-    pub fn live_files(&self) -> Result<Vec<LiveFile>, Error> {
-        match &self.storage {
-            Storage::Rocks(rocks) => rocks.underlying.live_files(),
-            _ => Ok(vec![]),
-        }
-    }
-}
-
-fn rocks_cf<'a>(rocks_db: &'a RocksDB, cf_name: &str) -> Arc<rocksdb::BoundColumnFamily<'a>> {
-    rocks_db
-        .underlying
-        .cf_handle(cf_name)
-        .expect("Map-keying column family should have been checked at DB creation")
-}
-
-fn rocks_cf_from_db<'a>(
-    db: &'a Database,
-    cf_name: &str,
-) -> Result<Arc<rocksdb::BoundColumnFamily<'a>>, TypedStoreError> {
-    match &db.storage {
-        Storage::Rocks(rocksdb) => Ok(rocksdb
+    fn write_batch(
+        &self,
+        batch: rocksdb::WriteBatch,
+        writeopts: &RocksWriteOptions,
+    ) -> Result<(), TypedStoreError> {
+        fail_point!("batch-write-before");
+        let ret = self
             .underlying
-            .cf_handle(cf_name)
-            .expect("Map-keying column family should have been checked at DB creation")),
-        _ => Err(TypedStoreError::RocksDB(
-            "using invalid batch type for the database".to_string(),
-        )),
+            .write_opt(batch, &writeopts.to_rocksdb())
+            .map_err(typed_store_err_from_rocks_err);
+        fail_point!("batch-write-after");
+        #[allow(clippy::let_and_return)]
+        ret
+    }
+
+    fn create_cf(&self, name: &str, opts: &rocksdb::Options) -> Result<(), TypedStoreError> {
+        self.underlying
+            .create_cf(name, opts)
+            .map_err(typed_store_err_from_rocks_err)
+    }
+
+    fn has_cf(&self, name: &str) -> bool {
+        self.underlying.cf_handle(name).is_some()
+    }
+
+    fn drop_cf(&self, name: &str) -> Result<(), TypedStoreError> {
+        self.underlying
+            .drop_cf(name)
+            .map_err(typed_store_err_from_rocks_err)
+    }
+
+    fn flush(&self) -> Result<(), TypedStoreError> {
+        self.underlying
+            .flush()
+            .map_err(typed_store_err_from_rocks_err)
+    }
+
+    fn checkpoint(&self, path: &Path) -> Result<(), TypedStoreError> {
+        let checkpoint =
+            Checkpoint::new(&self.underlying).map_err(typed_store_err_from_rocks_err)?;
+        checkpoint
+            .create_checkpoint(path)
+            .map_err(|e| TypedStoreError::RocksDB(e.to_string()))
+    }
+
+    fn compact_range(&self, cf_name: &str, start: Option<&[u8]>, end: Option<&[u8]>) {
+        self.underlying
+            .compact_range_cf(&self.rocks_cf(cf_name), start, end);
+    }
+
+    fn raw_iterator<'a>(&'a self, cf_name: &str, readopts: RocksReadOptions) -> RawIter<'a> {
+        self.underlying
+            .raw_iterator_cf_opt(&self.rocks_cf(cf_name), readopts.to_rocksdb())
+    }
+
+    fn key_may_exist(&self, cf_name: &str, key: &[u8], readopts: &RocksReadOptions) -> bool {
+        self.underlying
+            .key_may_exist_cf_opt(&self.rocks_cf(cf_name), key, &readopts.to_rocksdb())
+    }
+
+    fn try_catch_up_with_primary(&self) -> Result<(), TypedStoreError> {
+        self.underlying
+            .try_catch_up_with_primary()
+            .map_err(typed_store_err_from_rocks_err)
+    }
+
+    fn report_cf_metrics(&self, cf_name: &str, db_name: &str, db_metrics: &Arc<Self::Metrics>) {
+        let Some(cf) = self.underlying.cf_handle(cf_name) else {
+            tracing::warn!(
+                "unable to report metrics for cf {cf_name:?} in db {db_name:?}",
+            );
+            return;
+        };
+
+        const METRICS_ERROR: i64 = -1;
+
+        db_metrics
+            .cf_metrics
+            .rocksdb_total_sst_files_size
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::TOTAL_SST_FILES_SIZE)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_total_blob_files_size
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, ROCKSDB_PROPERTY_TOTAL_BLOB_FILES_SIZE)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        let total_num_files: i64 = (0..=6)
+            .map(|level| {
+                self.get_int_property(&cf, &num_files_at_level(level))
+                    .unwrap_or(METRICS_ERROR)
+            })
+            .sum();
+        db_metrics
+            .cf_metrics
+            .rocksdb_total_num_files
+            .with_label_values(&[cf_name])
+            .set(total_num_files);
+        db_metrics
+            .cf_metrics
+            .rocksdb_num_level0_files
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, &num_files_at_level(0))
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_current_size_active_mem_tables
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::CUR_SIZE_ACTIVE_MEM_TABLE)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_size_all_mem_tables
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::SIZE_ALL_MEM_TABLES)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_num_snapshots
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::NUM_SNAPSHOTS)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_oldest_snapshot_time
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::OLDEST_SNAPSHOT_TIME)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_actual_delayed_write_rate
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::ACTUAL_DELAYED_WRITE_RATE)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_is_write_stopped
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::IS_WRITE_STOPPED)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_block_cache_capacity
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::BLOCK_CACHE_CAPACITY)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_block_cache_usage
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::BLOCK_CACHE_USAGE)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_block_cache_pinned_usage
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::BLOCK_CACHE_PINNED_USAGE)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_estimate_table_readers_mem
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::ESTIMATE_TABLE_READERS_MEM)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_estimated_num_keys
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::ESTIMATE_NUM_KEYS)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_num_immutable_mem_tables
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::NUM_IMMUTABLE_MEM_TABLE)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_mem_table_flush_pending
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::MEM_TABLE_FLUSH_PENDING)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_compaction_pending
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::COMPACTION_PENDING)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_estimate_pending_compaction_bytes
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::ESTIMATE_PENDING_COMPACTION_BYTES)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_num_running_compactions
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::NUM_RUNNING_COMPACTIONS)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_num_running_flushes
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::NUM_RUNNING_FLUSHES)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_estimate_oldest_key_time
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::ESTIMATE_OLDEST_KEY_TIME)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_background_errors
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::BACKGROUND_ERRORS)
+                    .unwrap_or(METRICS_ERROR),
+            );
+        db_metrics
+            .cf_metrics
+            .rocksdb_base_level
+            .with_label_values(&[cf_name])
+            .set(
+                self.get_int_property(&cf, properties::BASE_LEVEL)
+                    .unwrap_or(METRICS_ERROR),
+            );
     }
 }
 
-// Check if the database is corrupted, and if so, panic.
-// If the corrupted key is not set, we set it to [1].
+// ---------------------------------------------------------------------------
+// RocksDB-specific extensions on Database<RocksDB>
+// ---------------------------------------------------------------------------
+
+impl Database<RocksDB> {
+    /// Return the filesystem path (used for SST file pruning).
+    pub fn path_for_pruning(&self) -> &Path {
+        self.storage.underlying.path()
+    }
+
+    /// Return a list of live SST files.
+    pub fn live_files(&self) -> Result<Vec<LiveFile>, rocksdb::Error> {
+        self.storage.underlying.live_files()
+    }
+}
+
+// ---------------------------------------------------------------------------
+// RocksDB-specific DBMap impl — reopen using ReadWriteOptions
+// ---------------------------------------------------------------------------
+
+impl<K, V> DBMap<K, V, RocksDB> {
+    /// Reopens an open database as a typed map operating under a specific
+    /// column family. If no column family is passed, the default column
+    /// family is used.
+    ///
+    /// ```
+    /// use core::fmt::Error;
+    /// use std::sync::Arc;
+    ///
+    /// use prometheus::Registry;
+    /// use tempfile::tempdir;
+    /// use typed_store::rocks::{metrics::DBMetrics, *};
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), Error> {
+    ///     /// Open the DB with all needed column families first.
+    ///     let rocks = open_cf(
+    ///         tempdir().unwrap(),
+    ///         None,
+    ///         MetricConf::default(),
+    ///         &["First_CF", "Second_CF"],
+    ///     )
+    ///     .unwrap();
+    ///     /// Attach the column families to specific maps.
+    ///     let db_cf_1 = DBMap::<u32, u32>::reopen(
+    ///         &rocks,
+    ///         Some("First_CF"),
+    ///         &ReadWriteOptions::default(),
+    ///         false,
+    ///     )
+    ///     .expect("Failed to open storage");
+    ///     Ok(())
+    /// }
+    /// ```
+    #[instrument(level = "debug", skip(db), err)]
+    pub fn reopen(
+        db: &Arc<Database<RocksDB>>,
+        opt_cf: Option<&str>,
+        rw_options: &ReadWriteOptions,
+        is_deprecated: bool,
+    ) -> Result<Self, TypedStoreError> {
+        let cf_key = opt_cf.unwrap_or("default").to_owned();
+        let read_opts = RocksReadOptions {
+            ignore_range_deletions: rw_options.ignore_range_deletions,
+            lower_bound: None,
+            upper_bound: None,
+        };
+        let write_opts = RocksWriteOptions {
+            sync_to_disk: rw_options.sync_to_disk,
+        };
+        Ok(DBMap::new(db.clone(), read_opts, write_opts, &cf_key, is_deprecated))
+    }
+}
+
+// ---------------------------------------------------------------------------
+// Corruption helpers (RocksDB-specific, not part of StorageEngine)
+// ---------------------------------------------------------------------------
+
 pub fn check_and_mark_db_corruption(path: &Path) -> Result<(), String> {
     let db = rocksdb::DB::open_default(path).map_err(|e| e.to_string())?;
 
@@ -525,1276 +793,14 @@ pub fn check_and_mark_db_corruption(path: &Path) -> Result<(), String> {
     Ok(())
 }
 
-pub fn unmark_db_corruption(path: &Path) -> Result<(), Error> {
+pub fn unmark_db_corruption(path: &Path) -> Result<(), rocksdb::Error> {
     rocksdb::DB::open_default(path)?.put(DB_CORRUPTED_KEY, [0])
 }
 
-#[derive(Debug, Default)]
-pub struct MetricConf {
-    pub db_name: String,
-    pub read_sample_interval: SamplingInterval,
-    pub write_sample_interval: SamplingInterval,
-    pub iter_sample_interval: SamplingInterval,
-}
+// ---------------------------------------------------------------------------
+// DBOptions
+// ---------------------------------------------------------------------------
 
-impl MetricConf {
-    pub fn new(db_name: &str) -> Self {
-        if db_name.is_empty() {
-            error!("A meaningful db name should be used for metrics reporting.")
-        }
-        Self {
-            db_name: db_name.to_string(),
-            read_sample_interval: SamplingInterval::default(),
-            write_sample_interval: SamplingInterval::default(),
-            iter_sample_interval: SamplingInterval::default(),
-        }
-    }
-
-    pub fn with_sampling(self, read_interval: SamplingInterval) -> Self {
-        Self {
-            db_name: self.db_name,
-            read_sample_interval: read_interval,
-            write_sample_interval: SamplingInterval::default(),
-            iter_sample_interval: SamplingInterval::default(),
-        }
-    }
-}
-const CF_METRICS_REPORT_PERIOD_SECS: u64 = 30;
-const METRICS_ERROR: i64 = -1;
-
-/// An interface to a rocksDB database, keyed by a columnfamily
-#[derive(Clone, Debug)]
-pub struct DBMap<K, V> {
-    pub db: Arc<Database>,
-    _phantom: PhantomData<fn(K) -> V>,
-    column_family: ColumnFamily,
-    // the column family under which the map is stored
-    cf: String,
-    pub opts: ReadWriteOptions,
-    db_metrics: Arc<DBMetrics>,
-    get_sample_interval: SamplingInterval,
-    multiget_sample_interval: SamplingInterval,
-    write_sample_interval: SamplingInterval,
-    iter_sample_interval: SamplingInterval,
-    _metrics_task_cancel_handle: Arc<oneshot::Sender<()>>,
-}
-
-unsafe impl<K: Send, V: Send> Send for DBMap<K, V> {}
-
-impl<K, V> DBMap<K, V> {
-    pub(crate) fn new(
-        db: Arc<Database>,
-        opts: &ReadWriteOptions,
-        opt_cf: &str,
-        column_family: ColumnFamily,
-        is_deprecated: bool,
-    ) -> Self {
-        let db_cloned = db.clone();
-        let db_metrics = DBMetrics::get();
-        let db_metrics_cloned = db_metrics.clone();
-        let cf = opt_cf.to_string();
-
-        let (sender, mut recv) = tokio::sync::oneshot::channel();
-        if !is_deprecated && matches!(db.storage, Storage::Rocks(_)) {
-            tokio::task::spawn(async move {
-                let mut interval =
-                    tokio::time::interval(Duration::from_secs(CF_METRICS_REPORT_PERIOD_SECS));
-                loop {
-                    tokio::select! {
-                        _ = interval.tick() => {
-                            let db = db_cloned.clone();
-                            let cf = cf.clone();
-                            let db_metrics = db_metrics.clone();
-                            if let Err(e) = tokio::task::spawn_blocking(move || {
-                                Self::report_rocksdb_metrics(&db, &cf, &db_metrics);
-                            }).await {
-                                error!("Failed to log metrics with error: {}", e);
-                            }
-                        }
-                        _ = &mut recv => break,
-                    }
-                }
-                debug!("Returning the cf metric logging task for DBMap: {}", &cf);
-            });
-        }
-        DBMap {
-            db: db.clone(),
-            opts: opts.clone(),
-            _phantom: PhantomData,
-            column_family,
-            cf: opt_cf.to_string(),
-            db_metrics: db_metrics_cloned,
-            _metrics_task_cancel_handle: Arc::new(sender),
-            get_sample_interval: db.get_sampling_interval(),
-            multiget_sample_interval: db.multiget_sampling_interval(),
-            write_sample_interval: db.write_sampling_interval(),
-            iter_sample_interval: db.iter_sampling_interval(),
-        }
-    }
-
-    /// Reopens an open database as a typed map operating under a specific
-    /// column family. if no column family is passed, the default column
-    /// family is used.
-    ///
-    /// ```
-    /// use core::fmt::Error;
-    /// use std::sync::Arc;
-    ///
-    /// use prometheus::Registry;
-    /// use tempfile::tempdir;
-    /// use typed_store::{metrics::DBMetrics, rocks::*};
-    /// #[tokio::main]
-    /// async fn main() -> Result<(), Error> {
-    ///     /// Open the DB with all needed column families first.
-    ///     let rocks = open_cf(
-    ///         tempdir().unwrap(),
-    ///         None,
-    ///         MetricConf::default(),
-    ///         &["First_CF", "Second_CF"],
-    ///     )
-    ///     .unwrap();
-    ///     /// Attach the column families to specific maps.
-    ///     let db_cf_1 = DBMap::<u32, u32>::reopen(
-    ///         &rocks,
-    ///         Some("First_CF"),
-    ///         &ReadWriteOptions::default(),
-    ///         false,
-    ///     )
-    ///     .expect("Failed to open storage");
-    ///     let db_cf_2 = DBMap::<u32, u32>::reopen(
-    ///         &rocks,
-    ///         Some("Second_CF"),
-    ///         &ReadWriteOptions::default(),
-    ///         false,
-    ///     )
-    ///     .expect("Failed to open storage");
-    ///     Ok(())
-    /// }
-    /// ```
-    #[instrument(level = "debug", skip(db), err)]
-    pub fn reopen(
-        db: &Arc<Database>,
-        opt_cf: Option<&str>,
-        rw_options: &ReadWriteOptions,
-        is_deprecated: bool,
-    ) -> Result<Self, TypedStoreError> {
-        let cf_key = opt_cf
-            .unwrap_or(rocksdb::DEFAULT_COLUMN_FAMILY_NAME)
-            .to_owned();
-
-        Ok(DBMap::new(
-            db.clone(),
-            rw_options,
-            &cf_key,
-            ColumnFamily::Rocks(cf_key.to_string()),
-            is_deprecated,
-        ))
-    }
-
-    pub fn cf_name(&self) -> &str {
-        &self.cf
-    }
-
-    pub fn batch(&self) -> DBBatch {
-        let batch = match &self.db.storage {
-            Storage::Rocks(_) => StorageWriteBatch::Rocks(WriteBatch::default()),
-            Storage::InMemory(_) => StorageWriteBatch::InMemory(InMemoryBatch::default()),
-        };
-        DBBatch::new(
-            &self.db,
-            batch,
-            self.opts.writeopts(),
-            &self.db_metrics,
-            &self.write_sample_interval,
-        )
-    }
-
-    pub fn compact_range<J: Serialize>(&self, start: &J, end: &J) -> Result<(), TypedStoreError> {
-        let from_buf = be_fix_int_ser(start)?;
-        let to_buf = be_fix_int_ser(end)?;
-        self.db
-            .compact_range_cf(&self.cf, Some(from_buf), Some(to_buf));
-        Ok(())
-    }
-
-    pub fn compact_range_raw(
-        &self,
-        cf_name: &str,
-        start: Vec<u8>,
-        end: Vec<u8>,
-    ) -> Result<(), TypedStoreError> {
-        self.db.compact_range_cf(cf_name, Some(start), Some(end));
-        Ok(())
-    }
-
-    /// Returns a vector of raw values corresponding to the keys provided.
-    fn multi_get_pinned<J>(
-        &self,
-        keys: impl IntoIterator<Item = J>,
-    ) -> Result<Vec<Option<GetResult<'_>>>, TypedStoreError>
-    where
-        J: Borrow<K>,
-        K: Serialize,
-    {
-        let _timer = self
-            .db_metrics
-            .op_metrics
-            .rocksdb_multiget_latency_seconds
-            .with_label_values(&[&self.cf])
-            .start_timer();
-        let perf_ctx = if self.multiget_sample_interval.sample() {
-            Some(RocksDBPerfContext)
-        } else {
-            None
-        };
-        let keys_bytes: Result<Vec<_>, TypedStoreError> = keys
-            .into_iter()
-            .map(|k| be_fix_int_ser(k.borrow()))
-            .collect();
-        let results: Result<Vec<_>, TypedStoreError> = self
-            .db
-            .multi_get(&self.column_family, keys_bytes?, &self.opts.readopts())
-            .into_iter()
-            .collect();
-        let entries = results?;
-        let entry_size = entries
-            .iter()
-            .flatten()
-            .map(|entry| entry.len())
-            .sum::<usize>();
-        self.db_metrics
-            .op_metrics
-            .rocksdb_multiget_bytes
-            .with_label_values(&[&self.cf])
-            .observe(entry_size as f64);
-        if perf_ctx.is_some() {
-            self.db_metrics
-                .read_perf_ctx_metrics
-                .report_metrics(&self.cf);
-        }
-        Ok(entries)
-    }
-
-    fn get_rocksdb_int_property(
-        rocksdb: &RocksDB,
-        cf: &impl AsColumnFamilyRef,
-        property_name: &std::ffi::CStr,
-    ) -> Result<i64, TypedStoreError> {
-        match rocksdb.underlying.property_int_value_cf(cf, property_name) {
-            Ok(Some(value)) => Ok(value.min(i64::MAX as u64).try_into().unwrap_or_default()),
-            Ok(None) => Ok(0),
-            Err(e) => Err(TypedStoreError::RocksDB(e.into_string())),
-        }
-    }
-
-    fn report_rocksdb_metrics(
-        database: &Arc<Database>,
-        cf_name: &str,
-        db_metrics: &Arc<DBMetrics>,
-    ) {
-        let Storage::Rocks(rocksdb) = &database.storage else {
-            return;
-        };
-
-        let Some(cf) = rocksdb.underlying.cf_handle(cf_name) else {
-            tracing::warn!(
-                "unable to report metrics for cf {cf_name:?} in db {:?}",
-                database.db_name()
-            );
-            return;
-        };
-
-        db_metrics
-            .cf_metrics
-            .rocksdb_total_sst_files_size
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::TOTAL_SST_FILES_SIZE)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_total_blob_files_size
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(
-                    rocksdb,
-                    &cf,
-                    ROCKSDB_PROPERTY_TOTAL_BLOB_FILES_SIZE,
-                )
-                .unwrap_or(METRICS_ERROR),
-            );
-        // 7 is the default number of levels in RocksDB. If we ever change the number of
-        // levels using `set_num_levels`, we need to update here as well. Note
-        // that there isn't an API to query the DB to get the number of levels (yet).
-        let total_num_files: i64 = (0..=6)
-            .map(|level| {
-                Self::get_rocksdb_int_property(rocksdb, &cf, &num_files_at_level(level))
-                    .unwrap_or(METRICS_ERROR)
-            })
-            .sum();
-        db_metrics
-            .cf_metrics
-            .rocksdb_total_num_files
-            .with_label_values(&[cf_name])
-            .set(total_num_files);
-        db_metrics
-            .cf_metrics
-            .rocksdb_num_level0_files
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, &num_files_at_level(0))
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_current_size_active_mem_tables
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::CUR_SIZE_ACTIVE_MEM_TABLE)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_size_all_mem_tables
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::SIZE_ALL_MEM_TABLES)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_num_snapshots
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::NUM_SNAPSHOTS)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_oldest_snapshot_time
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::OLDEST_SNAPSHOT_TIME)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_actual_delayed_write_rate
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::ACTUAL_DELAYED_WRITE_RATE)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_is_write_stopped
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::IS_WRITE_STOPPED)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_block_cache_capacity
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::BLOCK_CACHE_CAPACITY)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_block_cache_usage
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::BLOCK_CACHE_USAGE)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_block_cache_pinned_usage
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::BLOCK_CACHE_PINNED_USAGE)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_estimate_table_readers_mem
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(
-                    rocksdb,
-                    &cf,
-                    properties::ESTIMATE_TABLE_READERS_MEM,
-                )
-                .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_estimated_num_keys
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::ESTIMATE_NUM_KEYS)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_num_immutable_mem_tables
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::NUM_IMMUTABLE_MEM_TABLE)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_mem_table_flush_pending
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::MEM_TABLE_FLUSH_PENDING)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_compaction_pending
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::COMPACTION_PENDING)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_estimate_pending_compaction_bytes
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(
-                    rocksdb,
-                    &cf,
-                    properties::ESTIMATE_PENDING_COMPACTION_BYTES,
-                )
-                .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_num_running_compactions
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::NUM_RUNNING_COMPACTIONS)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_num_running_flushes
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::NUM_RUNNING_FLUSHES)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_estimate_oldest_key_time
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::ESTIMATE_OLDEST_KEY_TIME)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_background_errors
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::BACKGROUND_ERRORS)
-                    .unwrap_or(METRICS_ERROR),
-            );
-        db_metrics
-            .cf_metrics
-            .rocksdb_base_level
-            .with_label_values(&[cf_name])
-            .set(
-                Self::get_rocksdb_int_property(rocksdb, &cf, properties::BASE_LEVEL)
-                    .unwrap_or(METRICS_ERROR),
-            );
-    }
-
-    pub fn checkpoint_db(&self, path: &Path) -> Result<(), TypedStoreError> {
-        self.db.checkpoint(path)
-    }
-
-    pub fn table_summary(&self) -> eyre::Result<TableSummary>
-    where
-        K: Serialize + DeserializeOwned,
-        V: Serialize + DeserializeOwned,
-    {
-        let mut num_keys = 0;
-        let mut key_bytes_total = 0;
-        let mut value_bytes_total = 0;
-        let mut key_hist = hdrhistogram::Histogram::<u64>::new_with_max(100000, 2).unwrap();
-        let mut value_hist = hdrhistogram::Histogram::<u64>::new_with_max(100000, 2).unwrap();
-        for item in self.safe_iter() {
-            let (key, value) = item?;
-            num_keys += 1;
-            let key_len = be_fix_int_ser(key.borrow())?.len();
-            let value_len = bcs::to_bytes(value.borrow())?.len();
-            key_bytes_total += key_len;
-            value_bytes_total += value_len;
-            key_hist.record(key_len as u64)?;
-            value_hist.record(value_len as u64)?;
-        }
-        Ok(TableSummary {
-            num_keys,
-            key_bytes_total,
-            value_bytes_total,
-            key_hist,
-            value_hist,
-        })
-    }
-
-    // Creates metrics and context for tracking an iterator usage and performance.
-    fn create_iter_context(
-        &self,
-    ) -> (
-        Option<HistogramTimer>,
-        Option<Histogram>,
-        Option<Histogram>,
-        Option<RocksDBPerfContext>,
-    ) {
-        let timer = self
-            .db_metrics
-            .op_metrics
-            .rocksdb_iter_latency_seconds
-            .with_label_values(&[&self.cf])
-            .start_timer();
-        let bytes_scanned = self
-            .db_metrics
-            .op_metrics
-            .rocksdb_iter_bytes
-            .with_label_values(&[&self.cf]);
-        let keys_scanned = self
-            .db_metrics
-            .op_metrics
-            .rocksdb_iter_keys
-            .with_label_values(&[&self.cf]);
-        let perf_ctx = if self.iter_sample_interval.sample() {
-            Some(RocksDBPerfContext)
-        } else {
-            None
-        };
-        (
-            Some(timer),
-            Some(bytes_scanned),
-            Some(keys_scanned),
-            perf_ctx,
-        )
-    }
-
-    // Creates a RocksDB read option with specified lower and upper bounds.
-    /// Lower bound is inclusive, and upper bound is exclusive.
-    fn create_read_options_with_bounds(
-        &self,
-        lower_bound: Option<K>,
-        upper_bound: Option<K>,
-    ) -> ReadOptions
-    where
-        K: Serialize,
-    {
-        let mut readopts = self.opts.readopts();
-        if let Some(lower_bound) = lower_bound {
-            let key_buf = be_fix_int_ser(&lower_bound).unwrap();
-            readopts.set_iterate_lower_bound(key_buf);
-        }
-        if let Some(upper_bound) = upper_bound {
-            let key_buf = be_fix_int_ser(&upper_bound).unwrap();
-            readopts.set_iterate_upper_bound(key_buf);
-        }
-        readopts
-    }
-
-    /// Creates a safe reversed iterator with optional bounds.
-    /// Upper bound is included.
-    pub fn reversed_safe_iter_with_bounds(
-        &self,
-        lower_bound: Option<K>,
-        upper_bound: Option<K>,
-    ) -> Result<SafeRevIter<'_, K, V>, TypedStoreError>
-    where
-        K: Serialize + DeserializeOwned,
-        V: Serialize + DeserializeOwned,
-    {
-        let upper_bound_key = upper_bound.as_ref().map(|k| be_fix_int_ser(&k));
-        let readopts = self.create_read_options_with_range((
-            lower_bound
-                .as_ref()
-                .map(Bound::Included)
-                .unwrap_or(Bound::Unbounded),
-            upper_bound
-                .as_ref()
-                .map(Bound::Included)
-                .unwrap_or(Bound::Unbounded),
-        ));
-
-        let db_iter = self.db.raw_iterator_cf(&self.cf, readopts);
-        let (_timer, bytes_scanned, keys_scanned, _perf_ctx) = self.create_iter_context();
-        let iter = SafeIter::new(
-            self.cf.clone(),
-            db_iter,
-            _timer,
-            _perf_ctx,
-            bytes_scanned,
-            keys_scanned,
-            Some(self.db_metrics.clone()),
-        );
-        Ok(SafeRevIter::new(iter, upper_bound_key.transpose()?))
-    }
-
-    // Creates a RocksDB read option with lower and upper bounds set corresponding
-    // to `range`.
-    fn create_read_options_with_range(&self, range: impl RangeBounds<K>) -> ReadOptions
-    where
-        K: Serialize,
-    {
-        let mut readopts = self.opts.readopts();
-
-        let lower_bound = range.start_bound();
-        let upper_bound = range.end_bound();
-
-        match lower_bound {
-            Bound::Included(lower_bound) => {
-                // Rocksdb lower bound is inclusive by default so nothing to do
-                let key_buf = be_fix_int_ser(&lower_bound).expect("Serialization must not fail");
-                readopts.set_iterate_lower_bound(key_buf);
-            }
-            Bound::Excluded(lower_bound) => {
-                let mut key_buf =
-                    be_fix_int_ser(&lower_bound).expect("Serialization must not fail");
-
-                // Since we want exclusive, we need to increment the key to exclude the previous
-                big_endian_saturating_add_one(&mut key_buf);
-                readopts.set_iterate_lower_bound(key_buf);
-            }
-            Bound::Unbounded => (),
-        };
-
-        match upper_bound {
-            Bound::Included(upper_bound) => {
-                let mut key_buf =
-                    be_fix_int_ser(&upper_bound).expect("Serialization must not fail");
-
-                // If the key is already at the limit, there's nowhere else to go, so no upper
-                // bound
-                if !is_max(&key_buf) {
-                    // Since we want exclusive, we need to increment the key to get the upper bound
-                    big_endian_saturating_add_one(&mut key_buf);
-                    readopts.set_iterate_upper_bound(key_buf);
-                }
-            }
-            Bound::Excluded(upper_bound) => {
-                // Rocksdb upper bound is inclusive by default so nothing to do
-                let key_buf = be_fix_int_ser(&upper_bound).expect("Serialization must not fail");
-                readopts.set_iterate_upper_bound(key_buf);
-            }
-            Bound::Unbounded => (),
-        };
-
-        readopts
-    }
-}
-
-pub enum StorageWriteBatch {
-    Rocks(rocksdb::WriteBatch),
-    InMemory(InMemoryBatch),
-}
-
-/// Provides a mutable struct to form a collection of database write operations,
-/// and execute them.
-///
-/// Batching write and delete operations is faster than performing them one by
-/// one and ensures their atomicity,  ie. they are all written or none is.
-/// This is also true of operations across column families in the same database.
-///
-/// Serializations / Deserialization, and naming of column families is performed
-/// by passing a DBMap<K,V> with each operation.
-///
-/// ```
-/// use core::fmt::Error;
-/// use std::sync::Arc;
-///
-/// use prometheus::Registry;
-/// use tempfile::tempdir;
-/// use typed_store::{Map, metrics::DBMetrics, rocks::*};
-///
-/// #[tokio::main]
-/// async fn main() -> Result<(), Error> {
-///     let rocks = open_cf(
-///         tempfile::tempdir().unwrap(),
-///         None,
-///         MetricConf::default(),
-///         &["First_CF", "Second_CF"],
-///     )
-///     .unwrap();
-///
-///     let db_cf_1 = DBMap::reopen(
-///         &rocks,
-///         Some("First_CF"),
-///         &ReadWriteOptions::default(),
-///         false,
-///     )
-///     .expect("Failed to open storage");
-///     let keys_vals_1 = (1..100).map(|i| (i, i.to_string()));
-///
-///     let db_cf_2 = DBMap::reopen(
-///         &rocks,
-///         Some("Second_CF"),
-///         &ReadWriteOptions::default(),
-///         false,
-///     )
-///     .expect("Failed to open storage");
-///     let keys_vals_2 = (1000..1100).map(|i| (i, i.to_string()));
-///
-///     let mut batch = db_cf_1.batch();
-///     batch
-///         .insert_batch(&db_cf_1, keys_vals_1.clone())
-///         .expect("Failed to batch insert")
-///         .insert_batch(&db_cf_2, keys_vals_2.clone())
-///         .expect("Failed to batch insert");
-///
-///     let _ = batch.write().expect("Failed to execute batch");
-///     for (k, v) in keys_vals_1 {
-///         let val = db_cf_1.get(&k).expect("Failed to get inserted key");
-///         assert_eq!(Some(v), val);
-///     }
-///
-///     for (k, v) in keys_vals_2 {
-///         let val = db_cf_2.get(&k).expect("Failed to get inserted key");
-///         assert_eq!(Some(v), val);
-///     }
-///     Ok(())
-/// }
-/// ```
-pub struct DBBatch {
-    database: Arc<Database>,
-    batch: StorageWriteBatch,
-    opts: WriteOptions,
-    db_metrics: Arc<DBMetrics>,
-    write_sample_interval: SamplingInterval,
-}
-
-impl DBBatch {
-    /// Create a new batch associated with a DB reference.
-    ///
-    /// Use `open_cf` to get the DB reference or an existing open database.
-    pub fn new(
-        dbref: &Arc<Database>,
-        batch: StorageWriteBatch,
-        opts: WriteOptions,
-        db_metrics: &Arc<DBMetrics>,
-        write_sample_interval: &SamplingInterval,
-    ) -> Self {
-        DBBatch {
-            database: dbref.clone(),
-            batch,
-            opts,
-            db_metrics: db_metrics.clone(),
-            write_sample_interval: write_sample_interval.clone(),
-        }
-    }
-
-    /// Consume the batch and write its operations to the database
-    #[instrument(level = "trace", skip_all, err)]
-    pub fn write(self) -> Result<(), TypedStoreError> {
-        let db_name = self.database.db_name();
-        let timer = self
-            .db_metrics
-            .op_metrics
-            .rocksdb_batch_commit_latency_seconds
-            .with_label_values(&[&db_name])
-            .start_timer();
-        let batch_size = self.size_in_bytes();
-
-        let perf_ctx = if self.write_sample_interval.sample() {
-            Some(RocksDBPerfContext)
-        } else {
-            None
-        };
-        self.database.write(self.batch, &self.opts)?;
-        self.db_metrics
-            .op_metrics
-            .rocksdb_batch_commit_bytes
-            .with_label_values(&[&db_name])
-            .observe(batch_size as f64);
-
-        if perf_ctx.is_some() {
-            self.db_metrics
-                .write_perf_ctx_metrics
-                .report_metrics(&db_name);
-        }
-        let elapsed = timer.stop_and_record();
-        if elapsed > 1.0 {
-            warn!(?elapsed, ?db_name, "very slow batch write");
-            self.db_metrics
-                .op_metrics
-                .rocksdb_very_slow_batch_writes_count
-                .with_label_values(&[&db_name])
-                .inc();
-            self.db_metrics
-                .op_metrics
-                .rocksdb_very_slow_batch_writes_duration_ms
-                .with_label_values(&[&db_name])
-                .inc_by((elapsed * 1000.0) as u64);
-        }
-        Ok(())
-    }
-
-    pub fn size_in_bytes(&self) -> usize {
-        match self.batch {
-            StorageWriteBatch::Rocks(ref b) => b.size_in_bytes(),
-            StorageWriteBatch::InMemory(_) => 0,
-        }
-    }
-
-    pub fn delete_batch<J: Borrow<K>, K: Serialize, V>(
-        &mut self,
-        db: &DBMap<K, V>,
-        purged_vals: impl IntoIterator<Item = J>,
-    ) -> Result<(), TypedStoreError> {
-        if !Arc::ptr_eq(&db.db, &self.database) {
-            return Err(TypedStoreError::CrossDBBatch);
-        }
-
-        purged_vals
-            .into_iter()
-            .try_for_each::<_, Result<_, TypedStoreError>>(|k| {
-                let k_buf = be_fix_int_ser(k.borrow())?;
-                match (&mut self.batch, &db.column_family) {
-                    (StorageWriteBatch::Rocks(b), ColumnFamily::Rocks(name)) => {
-                        b.delete_cf(&rocks_cf_from_db(&self.database, name)?, k_buf)
-                    }
-                    (StorageWriteBatch::InMemory(b), ColumnFamily::InMemory(name)) => {
-                        b.delete_cf(name, k_buf)
-                    }
-                    _ => Err(TypedStoreError::RocksDB(
-                        "typed store invariant violation".to_string(),
-                    ))?,
-                }
-                Ok(())
-            })?;
-        Ok(())
-    }
-
-    /// Deletes a range of keys between `from` (inclusive) and `to`
-    /// (non-inclusive) by writing a range delete tombstone in the db map
-    /// If the DBMap is configured with ignore_range_deletions set to false,
-    /// the effect of this write will be visible immediately i.e. you won't
-    /// see old values when you do a lookup or scan. But if it is configured
-    /// with ignore_range_deletions set to true, the old value are visible until
-    /// compaction actually deletes them which will happen sometime after. By
-    /// default ignore_range_deletions is set to true on a DBMap (unless it is
-    /// overridden in the config), so please use this function with caution
-    pub fn schedule_delete_range<K: Serialize, V>(
-        &mut self,
-        db: &DBMap<K, V>,
-        from: &K,
-        to: &K,
-    ) -> Result<(), TypedStoreError> {
-        if !Arc::ptr_eq(&db.db, &self.database) {
-            return Err(TypedStoreError::CrossDBBatch);
-        }
-
-        let from_buf = be_fix_int_ser(from)?;
-        let to_buf = be_fix_int_ser(to)?;
-
-        if let StorageWriteBatch::Rocks(b) = &mut self.batch {
-            b.delete_range_cf(
-                &rocks_cf_from_db(&self.database, db.cf_name())?,
-                from_buf,
-                to_buf,
-            );
-        }
-        Ok(())
-    }
-
-    /// inserts a range of (key, value) pairs given as an iterator
-    pub fn insert_batch<J: Borrow<K>, K: Serialize, U: Borrow<V>, V: Serialize>(
-        &mut self,
-        db: &DBMap<K, V>,
-        new_vals: impl IntoIterator<Item = (J, U)>,
-    ) -> Result<&mut Self, TypedStoreError> {
-        if !Arc::ptr_eq(&db.db, &self.database) {
-            return Err(TypedStoreError::CrossDBBatch);
-        }
-        let mut total = 0usize;
-        new_vals
-            .into_iter()
-            .try_for_each::<_, Result<_, TypedStoreError>>(|(k, v)| {
-                let k_buf = be_fix_int_ser(k.borrow())?;
-                let v_buf = bcs::to_bytes(v.borrow()).map_err(typed_store_err_from_bcs_err)?;
-                total += k_buf.len() + v_buf.len();
-                match (&mut self.batch, &db.column_family) {
-                    (StorageWriteBatch::Rocks(b), ColumnFamily::Rocks(name)) => {
-                        b.put_cf(&rocks_cf_from_db(&self.database, name)?, k_buf, v_buf)
-                    }
-                    (StorageWriteBatch::InMemory(b), ColumnFamily::InMemory(name)) => {
-                        b.put_cf(name, k_buf, v_buf)
-                    }
-                    _ => Err(TypedStoreError::RocksDB(
-                        "typed store invariant violation".to_string(),
-                    ))?,
-                }
-                Ok(())
-            })?;
-        self.db_metrics
-            .op_metrics
-            .rocksdb_batch_put_bytes
-            .with_label_values(&[&db.cf])
-            .observe(total as f64);
-        Ok(self)
-    }
-}
-
-pub type RawIter<'a> = rocksdb::DBRawIteratorWithThreadMode<'a, DBWithThreadMode<MultiThreaded>>;
-
-impl<'a, K, V> Map<'a, K, V> for DBMap<K, V>
-where
-    K: Serialize + DeserializeOwned,
-    V: Serialize + DeserializeOwned,
-{
-    type Error = TypedStoreError;
-    type SafeIterator = SafeIter<'a, K, V>;
-
-    #[instrument(level = "trace", skip_all, err)]
-    fn contains_key(&self, key: &K) -> Result<bool, TypedStoreError> {
-        let key_buf = be_fix_int_ser(key)?;
-        let readopts = self.opts.readopts();
-        Ok(self.db.key_may_exist_cf(&self.cf, &key_buf, &readopts)
-            && self
-                .db
-                .get(&self.column_family, &key_buf, &readopts)?
-                .is_some())
-    }
-
-    #[instrument(level = "trace", skip_all, err)]
-    fn multi_contains_keys<J>(
-        &self,
-        keys: impl IntoIterator<Item = J>,
-    ) -> Result<Vec<bool>, Self::Error>
-    where
-        J: Borrow<K>,
-    {
-        let values = self.multi_get_pinned(keys)?;
-        Ok(values.into_iter().map(|v| v.is_some()).collect())
-    }
-
-    #[instrument(level = "trace", skip_all, err)]
-    fn get(&self, key: &K) -> Result<Option<V>, TypedStoreError> {
-        let _timer = self
-            .db_metrics
-            .op_metrics
-            .rocksdb_get_latency_seconds
-            .with_label_values(&[&self.cf])
-            .start_timer();
-        let perf_ctx = if self.get_sample_interval.sample() {
-            Some(RocksDBPerfContext)
-        } else {
-            None
-        };
-        let key_buf = be_fix_int_ser(key)?;
-        let res = self
-            .db
-            .get(&self.column_family, &key_buf, &self.opts.readopts())?;
-        self.db_metrics
-            .op_metrics
-            .rocksdb_get_bytes
-            .with_label_values(&[&self.cf])
-            .observe(res.as_ref().map_or(0.0, |v| v.len() as f64));
-        if perf_ctx.is_some() {
-            self.db_metrics
-                .read_perf_ctx_metrics
-                .report_metrics(&self.cf);
-        }
-        match res {
-            Some(data) => Ok(Some(
-                bcs::from_bytes(&data).map_err(typed_store_err_from_bcs_err)?,
-            )),
-            None => Ok(None),
-        }
-    }
-
-    #[instrument(level = "trace", skip_all, err)]
-    fn insert(&self, key: &K, value: &V) -> Result<(), TypedStoreError> {
-        let timer = self
-            .db_metrics
-            .op_metrics
-            .rocksdb_put_latency_seconds
-            .with_label_values(&[&self.cf])
-            .start_timer();
-        let perf_ctx = if self.write_sample_interval.sample() {
-            Some(RocksDBPerfContext)
-        } else {
-            None
-        };
-        let key_buf = be_fix_int_ser(key)?;
-        let value_buf = bcs::to_bytes(value).map_err(typed_store_err_from_bcs_err)?;
-        self.db_metrics
-            .op_metrics
-            .rocksdb_put_bytes
-            .with_label_values(&[&self.cf])
-            .observe((key_buf.len() + value_buf.len()) as f64);
-        if perf_ctx.is_some() {
-            self.db_metrics
-                .write_perf_ctx_metrics
-                .report_metrics(&self.cf);
-        }
-        self.db.put_cf(
-            &self.column_family,
-            key_buf,
-            value_buf,
-            &self.opts.writeopts(),
-        )?;
-
-        let elapsed = timer.stop_and_record();
-        if elapsed > 1.0 {
-            warn!(?elapsed, cf = ?self.cf, "very slow insert");
-            self.db_metrics
-                .op_metrics
-                .rocksdb_very_slow_puts_count
-                .with_label_values(&[&self.cf])
-                .inc();
-            self.db_metrics
-                .op_metrics
-                .rocksdb_very_slow_puts_duration_ms
-                .with_label_values(&[&self.cf])
-                .inc_by((elapsed * 1000.0) as u64);
-        }
-
-        Ok(())
-    }
-
-    #[instrument(level = "trace", skip_all, err)]
-    fn remove(&self, key: &K) -> Result<(), TypedStoreError> {
-        let _timer = self
-            .db_metrics
-            .op_metrics
-            .rocksdb_delete_latency_seconds
-            .with_label_values(&[&self.cf])
-            .start_timer();
-        let perf_ctx = if self.write_sample_interval.sample() {
-            Some(RocksDBPerfContext)
-        } else {
-            None
-        };
-        let key_buf = be_fix_int_ser(key)?;
-        self.db
-            .delete_cf(&self.column_family, key_buf, &self.opts.writeopts())?;
-        self.db_metrics
-            .op_metrics
-            .rocksdb_deletes
-            .with_label_values(&[&self.cf])
-            .inc();
-        if perf_ctx.is_some() {
-            self.db_metrics
-                .write_perf_ctx_metrics
-                .report_metrics(&self.cf);
-        }
-        Ok(())
-    }
-
-    /// This method first drops the existing column family and then creates a
-    /// new one with the same name. The two operations are not atomic and
-    /// hence it is possible to get into a race condition where the column
-    /// family has been dropped but new one is not created yet
-    #[instrument(level = "trace", skip_all, err)]
-    fn unsafe_clear(&self) -> Result<(), TypedStoreError> {
-        let _ = self.db.drop_cf(&self.cf);
-        self.db
-            .create_cf(self.cf.clone(), &default_db_options().options)
-            .map_err(typed_store_err_from_rocks_err)?;
-        Ok(())
-    }
-
-    /// Writes a range delete tombstone to delete all entries in the db map
-    /// If the DBMap is configured with ignore_range_deletions set to false,
-    /// the effect of this write will be visible immediately i.e. you won't
-    /// see old values when you do a lookup or scan. But if it is configured
-    /// with ignore_range_deletions set to true, the old value are visible until
-    /// compaction actually deletes them which will happen sometime after. By
-    /// default ignore_range_deletions is set to true on a DBMap (unless it is
-    /// overridden in the config), so please use this function with caution
-    #[instrument(level = "trace", skip_all, err)]
-    fn schedule_delete_all(&self) -> Result<(), TypedStoreError> {
-        let first_key = self.safe_iter().next().transpose()?.map(|(k, _v)| k);
-        let last_key = self
-            .reversed_safe_iter_with_bounds(None, None)?
-            .next()
-            .transpose()?
-            .map(|(k, _v)| k);
-        if let Some((first_key, last_key)) = first_key.zip(last_key) {
-            let mut batch = self.batch();
-            batch.schedule_delete_range(self, &first_key, &last_key)?;
-            batch.write()?;
-        }
-        Ok(())
-    }
-
-    fn is_empty(&self) -> bool {
-        self.safe_iter().next().is_none()
-    }
-
-    fn safe_iter(&'a self) -> Self::SafeIterator {
-        let db_iter = self.db.raw_iterator_cf(&self.cf, self.opts.readopts());
-        let (_timer, bytes_scanned, keys_scanned, _perf_ctx) = self.create_iter_context();
-        SafeIter::new(
-            self.cf.clone(),
-            db_iter,
-            _timer,
-            _perf_ctx,
-            bytes_scanned,
-            keys_scanned,
-            Some(self.db_metrics.clone()),
-        )
-    }
-
-    fn safe_iter_with_bounds(
-        &'a self,
-        lower_bound: Option<K>,
-        upper_bound: Option<K>,
-    ) -> Self::SafeIterator {
-        let readopts = self.create_read_options_with_bounds(lower_bound, upper_bound);
-        let db_iter = self.db.raw_iterator_cf(&self.cf, readopts);
-        let (_timer, bytes_scanned, keys_scanned, _perf_ctx) = self.create_iter_context();
-        SafeIter::new(
-            self.cf.clone(),
-            db_iter,
-            _timer,
-            _perf_ctx,
-            bytes_scanned,
-            keys_scanned,
-            Some(self.db_metrics.clone()),
-        )
-    }
-
-    fn safe_range_iter(&'a self, range: impl RangeBounds<K>) -> Self::SafeIterator {
-        let readopts = self.create_read_options_with_range(range);
-        let db_iter = self.db.raw_iterator_cf(&self.cf, readopts);
-        let (_timer, bytes_scanned, keys_scanned, _perf_ctx) = self.create_iter_context();
-        SafeIter::new(
-            self.cf.clone(),
-            db_iter,
-            _timer,
-            _perf_ctx,
-            bytes_scanned,
-            keys_scanned,
-            Some(self.db_metrics.clone()),
-        )
-    }
-
-    /// Returns a vector of values corresponding to the keys provided.
-    #[instrument(level = "trace", skip_all, err)]
-    fn multi_get<J>(
-        &self,
-        keys: impl IntoIterator<Item = J>,
-    ) -> Result<Vec<Option<V>>, TypedStoreError>
-    where
-        J: Borrow<K>,
-    {
-        let results = self.multi_get_pinned(keys)?;
-        let values_parsed: Result<Vec<_>, TypedStoreError> = results
-            .into_iter()
-            .map(|value_byte| match value_byte {
-                Some(data) => Ok(Some(
-                    bcs::from_bytes(&data).map_err(typed_store_err_from_bcs_err)?,
-                )),
-                None => Ok(None),
-            })
-            .collect();
-
-        values_parsed
-    }
-
-    /// Convenience method for batch insertion
-    #[instrument(level = "trace", skip_all, err)]
-    fn multi_insert<J, U>(
-        &self,
-        key_val_pairs: impl IntoIterator<Item = (J, U)>,
-    ) -> Result<(), Self::Error>
-    where
-        J: Borrow<K>,
-        U: Borrow<V>,
-    {
-        let mut batch = self.batch();
-        batch.insert_batch(self, key_val_pairs)?;
-        batch.write()
-    }
-
-    /// Convenience method for batch removal
-    #[instrument(level = "trace", skip_all, err)]
-    fn multi_remove<J>(&self, keys: impl IntoIterator<Item = J>) -> Result<(), Self::Error>
-    where
-        J: Borrow<K>,
-    {
-        let mut batch = self.batch();
-        batch.delete_batch(self, keys)?;
-        batch.write()
-    }
-
-    /// Try to catch up with primary when running as secondary
-    #[instrument(level = "trace", skip_all, err)]
-    fn try_catch_up_with_primary(&self) -> Result<(), Self::Error> {
-        if let Storage::Rocks(rocks) = &self.db.storage {
-            rocks
-                .underlying
-                .try_catch_up_with_primary()
-                .map_err(typed_store_err_from_rocks_err)?;
-        }
-        Ok(())
-    }
-}
-
-pub fn read_size_from_env(var_name: &str) -> Option<usize> {
-    env::var(var_name)
-        .ok()?
-        .parse::<usize>()
-        .tap_err(|e| {
-            warn!(
-                "Env var {} does not contain valid usize integer: {}",
-                var_name, e
-            )
-        })
-        .ok()
-}
-
-#[derive(Clone, Debug)]
-pub struct ReadWriteOptions {
-    pub ignore_range_deletions: bool,
-    // Whether to sync to disk on every write.
-    sync_to_disk: bool,
-}
-
-impl ReadWriteOptions {
-    pub fn readopts(&self) -> ReadOptions {
-        let mut readopts = ReadOptions::default();
-        readopts.set_ignore_range_deletions(self.ignore_range_deletions);
-        readopts
-    }
-
-    pub fn writeopts(&self) -> WriteOptions {
-        let mut opts = WriteOptions::default();
-        opts.set_sync(self.sync_to_disk);
-        opts
-    }
-
-    pub fn set_ignore_range_deletions(mut self, ignore: bool) -> Self {
-        self.ignore_range_deletions = ignore;
-        self
-    }
-}
-
-impl Default for ReadWriteOptions {
-    fn default() -> Self {
-        Self {
-            ignore_range_deletions: true,
-            sync_to_disk: std::env::var("IOTA_DB_SYNC_TO_DISK").is_ok_and(|v| v != "0"),
-        }
-    }
-}
 // TODO: refactor this into a builder pattern, where rocksdb::Options are
 // generated after a call to build().
 #[derive(Default, Clone)]
@@ -1804,46 +810,32 @@ pub struct DBOptions {
 }
 
 impl DBOptions {
-    // Optimize lookup perf for tables where no scans are performed.
-    // If non-trivial number of values can be > 512B in size, it is beneficial to
-    // also specify optimize_for_large_values_no_scan().
     pub fn optimize_for_point_lookup(mut self, block_cache_size_mb: usize) -> DBOptions {
-        // NOTE: this overwrites the block options.
         self.options
             .optimize_for_point_lookup(block_cache_size_mb as u64);
         self
     }
 
-    // Optimize write and lookup perf for tables which are rarely scanned, and have
-    // large values. https://rocksdb.org/blog/2021/05/26/integrated-blob-db.html
     pub fn optimize_for_large_values_no_scan(mut self, min_blob_size: u64) -> DBOptions {
         if env::var(ENV_VAR_DISABLE_BLOB_STORAGE).is_ok() {
             info!("Large value blob storage optimization is disabled via env var.");
             return self;
         }
 
-        // Blob settings.
         self.options.set_enable_blob_files(true);
         self.options
             .set_blob_compression_type(rocksdb::DBCompressionType::Lz4);
         self.options.set_enable_blob_gc(true);
-        // Since each blob can have non-trivial size overhead, and compression does not
-        // work across blobs, set a min blob size in bytes to so small
-        // transactions and effects are kept in sst files.
         self.options.set_min_blob_size(min_blob_size);
 
-        // Increase write buffer size to 256MiB.
         let write_buffer_size = read_size_from_env(ENV_VAR_MAX_WRITE_BUFFER_SIZE_MB)
             .unwrap_or(DEFAULT_MAX_WRITE_BUFFER_SIZE_MB)
             * 1024
             * 1024;
         self.options.set_write_buffer_size(write_buffer_size);
-        // Since large blobs are not in sst files, reduce the target file size and base
-        // level target size.
         let target_file_size_base = 64 << 20;
         self.options
             .set_target_file_size_base(target_file_size_base);
-        // Level 1 default to 64MiB * 4 ~ 256MiB.
         let max_level_zero_file_num = read_size_from_env(ENV_VAR_L0_NUM_FILES_COMPACTION_TRIGGER)
             .unwrap_or(DEFAULT_L0_NUM_FILES_COMPACTION_TRIGGER);
         self.options
@@ -1852,14 +844,12 @@ impl DBOptions {
         self
     }
 
-    // Optimize tables with a mix of lookup and scan workloads.
     pub fn optimize_for_read(mut self, block_cache_size_mb: usize) -> DBOptions {
         self.options
             .set_block_based_table_factory(&get_block_options(block_cache_size_mb, 16 << 10));
         self
     }
 
-    // Optimize DB receiving significant insertions.
     pub fn optimize_db_for_write_throughput(mut self, db_max_write_buffer_gb: u64) -> DBOptions {
         self.options
             .set_db_write_buffer_size(db_max_write_buffer_gb as usize * 1024 * 1024 * 1024);
@@ -1868,24 +858,19 @@ impl DBOptions {
         self
     }
 
-    // Optimize tables receiving significant insertions.
     pub fn optimize_for_write_throughput(mut self) -> DBOptions {
-        // Increase write buffer size to 256MiB.
         let write_buffer_size = read_size_from_env(ENV_VAR_MAX_WRITE_BUFFER_SIZE_MB)
             .unwrap_or(DEFAULT_MAX_WRITE_BUFFER_SIZE_MB)
             * 1024
             * 1024;
         self.options.set_write_buffer_size(write_buffer_size);
-        // Increase write buffers to keep to 6 before slowing down writes.
         let max_write_buffer_number = read_size_from_env(ENV_VAR_MAX_WRITE_BUFFER_NUMBER)
             .unwrap_or(DEFAULT_MAX_WRITE_BUFFER_NUMBER);
         self.options
             .set_max_write_buffer_number(max_write_buffer_number.try_into().unwrap());
-        // Keep 1 write buffer so recent writes can be read from memory.
         self.options
             .set_max_write_buffer_size_to_maintain((write_buffer_size).try_into().unwrap());
 
-        // Increase compaction trigger for level 0 to 6.
         let max_level_zero_file_num = read_size_from_env(ENV_VAR_L0_NUM_FILES_COMPACTION_TRIGGER)
             .unwrap_or(DEFAULT_L0_NUM_FILES_COMPACTION_TRIGGER);
         self.options.set_level_zero_file_num_compaction_trigger(
@@ -1896,42 +881,31 @@ impl DBOptions {
         );
         self.options
             .set_level_zero_stop_writes_trigger((max_level_zero_file_num * 16).try_into().unwrap());
-
-        // Increase sst file size to 128MiB.
         self.options.set_target_file_size_base(
             read_size_from_env(ENV_VAR_TARGET_FILE_SIZE_BASE_MB)
                 .unwrap_or(DEFAULT_TARGET_FILE_SIZE_BASE_MB) as u64
                 * 1024
                 * 1024,
         );
-
-        // Increase level 1 target size to 256MiB * 6 ~ 1.5GiB.
         self.options
             .set_max_bytes_for_level_base((write_buffer_size * max_level_zero_file_num) as u64);
 
         self
     }
 
-    // Optimize tables receiving significant insertions, without any deletions.
-    // TODO: merge this function with optimize_for_write_throughput(), and use a
-    // flag to indicate if deletion is received.
     pub fn optimize_for_write_throughput_no_deletion(mut self) -> DBOptions {
-        // Increase write buffer size to 256MiB.
         let write_buffer_size = read_size_from_env(ENV_VAR_MAX_WRITE_BUFFER_SIZE_MB)
             .unwrap_or(DEFAULT_MAX_WRITE_BUFFER_SIZE_MB)
             * 1024
             * 1024;
         self.options.set_write_buffer_size(write_buffer_size);
-        // Increase write buffers to keep to 6 before slowing down writes.
         let max_write_buffer_number = read_size_from_env(ENV_VAR_MAX_WRITE_BUFFER_NUMBER)
             .unwrap_or(DEFAULT_MAX_WRITE_BUFFER_NUMBER);
         self.options
             .set_max_write_buffer_number(max_write_buffer_number.try_into().unwrap());
-        // Keep 1 write buffer so recent writes can be read from memory.
         self.options
             .set_max_write_buffer_size_to_maintain((write_buffer_size).try_into().unwrap());
 
-        // Switch to universal compactions.
         self.options
             .set_compaction_style(rocksdb::DBCompactionStyle::Universal);
         let mut compaction_options = rocksdb::UniversalCompactOptions::default();
@@ -1950,23 +924,18 @@ impl DBOptions {
         );
         self.options
             .set_level_zero_stop_writes_trigger((max_level_zero_file_num * 16).try_into().unwrap());
-
-        // Increase sst file size to 128MiB.
         self.options.set_target_file_size_base(
             read_size_from_env(ENV_VAR_TARGET_FILE_SIZE_BASE_MB)
                 .unwrap_or(DEFAULT_TARGET_FILE_SIZE_BASE_MB) as u64
                 * 1024
                 * 1024,
         );
-
-        // This should be a no-op for universal compaction but increasing it to be safe.
         self.options
             .set_max_bytes_for_level_base((write_buffer_size * max_level_zero_file_num) as u64);
 
         self
     }
 
-    // Overrides the block options with different block cache size and block size.
     pub fn set_block_options(
         mut self,
         block_cache_size_mb: usize,
@@ -1980,7 +949,6 @@ impl DBOptions {
         self
     }
 
-    // Disables write stalling and stopping based on pending compaction bytes.
     pub fn disable_write_throttling(mut self) -> DBOptions {
         self.options.set_soft_pending_compaction_bytes_limit(0);
         self.options.set_hard_pending_compaction_bytes_limit(0);
@@ -1993,34 +961,16 @@ impl DBOptions {
 pub fn default_db_options() -> DBOptions {
     let mut opt = rocksdb::Options::default();
 
-    // One common issue when running tests on Mac is that the default ulimit is too
-    // low, leading to I/O errors such as "Too many open files". Raising fdlimit
-    // to bypass it.
     if let Some(limit) = fdlimit::raise_fd_limit() {
-        // on windows raise_fd_limit return None
         opt.set_max_open_files((limit / 8) as i32);
     }
 
-    // The table cache is locked for updates and this determines the number
-    // of shards, ie 2^10. Increase in case of lock contentions.
     opt.set_table_cache_num_shard_bits(10);
 
-    // LSM compression settings
     opt.set_compression_type(rocksdb::DBCompressionType::Lz4);
     opt.set_bottommost_compression_type(rocksdb::DBCompressionType::Zstd);
     opt.set_bottommost_zstd_max_train_bytes(1024 * 1024, true);
 
-    // IOTA uses multiple RocksDB in a node, so total sizes of write buffers and WAL
-    // can be higher than the limits below.
-    //
-    // RocksDB also exposes the option to configure total write buffer size across
-    // multiple instances via `write_buffer_manager`. But the write buffer flush
-    // policy (flushing the buffer receiving the next write) may not work well.
-    // So sticking to per-db write buffer size limit for now.
-    //
-    // The environment variables are only meant to be emergency overrides. They may
-    // go away in future. It is preferable to update the default value, or
-    // override the option in code.
     opt.set_db_write_buffer_size(
         read_size_from_env(ENV_VAR_DB_WRITE_BUFFER_SIZE).unwrap_or(DEFAULT_DB_WRITE_BUFFER_SIZE)
             * 1024
@@ -2030,16 +980,12 @@ pub fn default_db_options() -> DBOptions {
         read_size_from_env(ENV_VAR_DB_WAL_SIZE).unwrap_or(DEFAULT_DB_WAL_SIZE) as u64 * 1024 * 1024,
     );
 
-    // Num threads for compactions and memtable flushes.
     opt.increase_parallelism(read_size_from_env(ENV_VAR_DB_PARALLELISM).unwrap_or(8) as i32);
 
     opt.set_enable_pipelined_write(true);
 
-    // Increase block size to 16KiB.
-    // https://github.com/EighteenZi/rocksdb_wiki/blob/master/Memory-usage-in-RocksDB.md#indexes-and-filter-blocks
     opt.set_block_based_table_factory(&get_block_options(128, 16 << 10));
 
-    // Set memtable bloomfilter.
     opt.set_memtable_prefix_bloom_ratio(0.02);
 
     DBOptions {
@@ -2049,20 +995,29 @@ pub fn default_db_options() -> DBOptions {
 }
 
 fn get_block_options(block_cache_size_mb: usize, block_size_bytes: usize) -> BlockBasedOptions {
-    // Set options mostly similar to those used in optimize_for_point_lookup(),
-    // except non-default binary and hash index, to hopefully reduce lookup
-    // latencies without causing any regression for scanning, with slightly more
-    // memory usages. https://github.com/facebook/rocksdb/blob/11cb6af6e5009c51794641905ca40ce5beec7fee/options/options.cc#L611-L621
     let mut block_options = BlockBasedOptions::default();
-    // Overrides block size.
     block_options.set_block_size(block_size_bytes);
-    // Configure a block cache.
     block_options.set_block_cache(&Cache::new_lru_cache(block_cache_size_mb << 20));
-    // Set a bloomfilter with 1% false positive rate.
     block_options.set_bloom_filter(10.0, false);
-    // From https://github.com/EighteenZi/rocksdb_wiki/blob/master/Block-Cache.md#caching-index-and-filter-blocks
     block_options.set_pin_l0_filter_and_index_blocks_in_cache(true);
     block_options
+}
+
+// ---------------------------------------------------------------------------
+// DB open helpers
+// ---------------------------------------------------------------------------
+
+pub fn read_size_from_env(var_name: &str) -> Option<usize> {
+    env::var(var_name)
+        .ok()?
+        .parse::<usize>()
+        .tap_err(|e| {
+            warn!(
+                "Env var {} does not contain valid usize integer: {}",
+                var_name, e
+            )
+        })
+        .ok()
 }
 
 /// Opens a database with options, and a number of column families that are
@@ -2088,7 +1043,6 @@ pub fn open_cf<P: AsRef<Path>>(
 }
 
 fn prepare_db_options(db_options: Option<rocksdb::Options>) -> rocksdb::Options {
-    // Customize database options
     let mut options = db_options.unwrap_or_else(|| default_db_options().options);
     options.create_if_missing(true);
     options.create_missing_column_families(true);
@@ -2105,15 +1059,6 @@ pub fn open_cf_opts<P: AsRef<Path>>(
     opt_cfs: &[(&str, rocksdb::Options)],
 ) -> Result<Arc<Database>, TypedStoreError> {
     let path = path.as_ref();
-    // In the simulator, we intercept the wall clock in the test thread only. This
-    // causes problems because rocksdb uses the simulated clock when creating
-    // its background threads, but then those threads see the real wall clock
-    // (because they are not the test thread), which causes rocksdb to panic.
-    // The `nondeterministic` macro evaluates expressions in new threads, which
-    // resolves the issue.
-    //
-    // This is a no-op in non-simulator builds.
-
     let cfs = populate_missing_cfs(opt_cfs, path).map_err(typed_store_err_from_rocks_err)?;
     nondeterministic!({
         let options = prepare_db_options(db_options);
@@ -2127,9 +1072,9 @@ pub fn open_cf_opts<P: AsRef<Path>>(
             .map_err(typed_store_err_from_rocks_err)?
         };
         Ok(Arc::new(Database::new(
-            Storage::Rocks(RocksDB {
+            RocksDB {
                 underlying: rocksdb,
-            }),
+            },
             metric_conf,
         )))
     })
@@ -2146,13 +1091,10 @@ pub fn open_cf_opts_secondary<P: AsRef<Path>>(
 ) -> Result<Arc<Database>, TypedStoreError> {
     let primary_path = primary_path.as_ref();
     let secondary_path = secondary_path.as_ref().map(|p| p.as_ref());
-    // See comment above for explanation of why nondeterministic is necessary here.
     nondeterministic!({
-        // Customize database options
         let mut options = db_options.unwrap_or_else(|| default_db_options().options);
 
         fdlimit::raise_fd_limit();
-        // This is a requirement by RocksDB when opening as secondary
         options.set_max_open_files(-1);
 
         let mut opt_cfs: std::collections::HashMap<_, _> = opt_cfs.iter().cloned().collect();
@@ -2161,7 +1103,6 @@ pub fn open_cf_opts_secondary<P: AsRef<Path>>(
             .unwrap_or_default();
 
         let default_db_options = default_db_options();
-        // Add CFs not explicitly listed
         for cf_key in cfs.iter() {
             if !opt_cfs.contains_key(&cf_key[..]) {
                 opt_cfs.insert(cf_key, default_db_options.options.clone());
@@ -2193,9 +1134,9 @@ pub fn open_cf_opts_secondary<P: AsRef<Path>>(
             db
         };
         Ok(Arc::new(Database::new(
-            Storage::Rocks(RocksDB {
+            RocksDB {
                 underlying: rocksdb,
-            }),
+            },
             metric_conf,
         )))
     })
@@ -2210,7 +1151,6 @@ pub fn list_tables(path: std::path::PathBuf) -> eyre::Result<Vec<String>> {
         .map(|q| {
             q.iter()
                 .filter_map(|s| {
-                    // The `default` table is not used
                     if s != DB_DEFAULT_CF_NAME {
                         Some(s.clone())
                     } else {
@@ -2221,27 +1161,14 @@ pub fn list_tables(path: std::path::PathBuf) -> eyre::Result<Vec<String>> {
         })
 }
 
-/// TODO: Good description of why we're doing this : RocksDB stores keys in BE and has a seek operator on iterators, see `https://github.com/facebook/rocksdb/wiki/Iterator#introduction`
-#[inline]
-pub fn be_fix_int_ser<S>(t: &S) -> Result<Vec<u8>, TypedStoreError>
-where
-    S: ?Sized + serde::Serialize,
-{
-    bincode::DefaultOptions::new()
-        .with_big_endian()
-        .with_fixint_encoding()
-        .serialize(t)
-        .map_err(typed_store_err_from_bincode_err)
-}
-
 #[derive(Clone)]
-pub struct DBMapTableConfigMap(BTreeMap<String, DBOptions>);
+pub struct DBMapTableConfigMap(std::collections::BTreeMap<String, DBOptions>);
 impl DBMapTableConfigMap {
-    pub fn new(map: BTreeMap<String, DBOptions>) -> Self {
+    pub fn new(map: std::collections::BTreeMap<String, DBOptions>) -> Self {
         Self(map)
     }
 
-    pub fn to_map(&self) -> BTreeMap<String, DBOptions> {
+    pub fn to_map(&self) -> std::collections::BTreeMap<String, DBOptions> {
         self.0.clone()
     }
 }
@@ -2251,7 +1178,6 @@ pub enum RocksDBAccessType {
     Secondary(Option<PathBuf>),
 }
 
-// Drops a database if there is no other handle to it, with retries and timeout.
 pub async fn safe_drop_db(path: PathBuf, timeout: Duration) -> Result<(), rocksdb::Error> {
     let mut backoff = backoff::ExponentialBackoff {
         max_elapsed_time: Some(timeout),
@@ -2290,57 +1216,4 @@ fn populate_missing_cfs(
             .map(|(name, opts)| (name.to_string(), (*opts).clone())),
     );
     Ok(cfs)
-}
-
-/// Given a `vec<u8>`, find the value which is one more than the vector
-/// if the vector was a big endian number.
-/// If the vector is already minimum, don't change it.
-fn big_endian_saturating_add_one(v: &mut [u8]) {
-    if is_max(v) {
-        return;
-    }
-    for i in (0..v.len()).rev() {
-        if v[i] == u8::MAX {
-            v[i] = 0;
-        } else {
-            v[i] += 1;
-            break;
-        }
-    }
-}
-
-/// Check if all the bytes in the vector are 0xFF
-fn is_max(v: &[u8]) -> bool {
-    v.iter().all(|&x| x == u8::MAX)
-}
-
-// `clippy::manual_div_ceil` is raised by code expanded by the
-// `uint::construct_uint!` macro so it needs to be fixed by `uint`
-#[expect(clippy::assign_op_pattern, clippy::manual_div_ceil)]
-#[test]
-fn test_helpers() {
-    let v = vec![];
-    assert!(is_max(&v));
-
-    fn check_add(v: Vec<u8>) {
-        let mut v = v;
-        let num = Num32::from_big_endian(&v);
-        big_endian_saturating_add_one(&mut v);
-        assert!(num + 1 == Num32::from_big_endian(&v));
-    }
-
-    uint::construct_uint! {
-        // 32 byte number
-        struct Num32(4);
-    }
-
-    let mut v = vec![255; 32];
-    big_endian_saturating_add_one(&mut v);
-    assert!(Num32::MAX == Num32::from_big_endian(&v));
-
-    check_add(vec![1; 32]);
-    check_add(vec![6; 32]);
-    check_add(vec![254; 32]);
-
-    // TBD: More tests coming with randomized arrays
 }
